@@ -7,10 +7,10 @@
         "@babel/preset-react": "^7.16.7",
         "axios": "^0.26.0",
         "bootstrap": "^5.1.3",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-bootstrap": "^2.1.2",
-        "react-dom": "^17.0.2",
-        "react-router-dom": "^6.2.1",
+        "react-dom": "^18.2.0",
+        "react-router-dom": "^6.10.0",
         "rest": "^2.0.0"
     },
     "scripts": {
